// coreui-gh/src/components/ui/data-table/types.ts
export type Option = {
    label: string
    value: string
    icon?: React.ComponentType<{ className?: string }>
}

export interface DataTableSearchableColumn<TData> {
    id: keyof TData
    title: string
}

export interface DataTableFilterableColumn<TData>
    extends DataTableSearchableColumn<TData> {
    options: Option[]
}

export interface DataTableFilterOption<TData> {
    id?: string
    label: string
    value: keyof TData | string
    items: Option[]
    isMulti?: boolean
}